Removes striping behavior from rows in legacy data grid when hovered.
This doesn't change the standard behavior, but in high contrast/forced
color modes this behavior was causing every other row to be black text
on a black background which was completely unreadable. This seems to be
the same reasoning that was used to add the not selected selector to this
rule as well in the past.

Bug: None
Change-Id: If018c15533e319a2cabcd35a7f1fa0c687085784
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3753276
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>