[Delegate-CH] (2) Add tests to devtools
This CL verifies that the issues sent by the accept-ch tag are still
dispatched by the delegate-ch tag. No functionality change should be
required for this to work in devtools.

This CL is part of a series:
(1) Add support to chromium
(2) Add tests to devtools

Bug: 1334152
Change-Id: I2e9948e97a83f2a28fa8530be514e260a355b7bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3703498
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>