[lighthouse] Fix lighthouse web tests
This CL fixes the lighthouse web tests sitting in the DevTools repo.
The `GetStorageAndQuota` CDP method can fail now due to storage
partitioning, so we have to account for that error.

Another bug seems to be that changing the throttling drop down
does not affect the underlying setting, so we change the test to
massage the settings directly instead of poking the UI.

Bug: None
Change-Id: Ib7f4e3845722bde7dd1fd86b99896a8339414d87
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3769874
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>