Toggle show ignore-listed call frames
This changes the way ignore-listing is disabled in the call stack.
In place of the link that previously allowed ignore-listing to be
turned off (but not back on), we now have a checkbox allowing the
ignored frames to be unhidden and then hidden again.

Bug: 1336604
Change-Id: Ie42dbe4884c571bf0a696c83c2695610f0ff53e8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3769849
Reviewed-by: Victor Porof <victorporof@chromium.org>
Commit-Queue: Victor Porof <victorporof@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>