Add a checkbox that controls automatic ignore listing
Third-party code can be automatically ignore-listed based on source map hints. Chrome DevTools has introduced a source map extension to identify third-party sources in source maps with the `x_google_ignoreList` field.

This CL adds the ability for the user to toggle this functionality on/off. It is enabled by default.

Bug: 1323199
Change-Id: Id35c74063af2526b4b27f60164ffe36f66a66a72
Signed-off-by: Victor Porof <victorporof@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3779791
Reviewed-by: Eric Leese <leese@chromium.org>