Prevent host bindings from loading resources from UNC paths
This CL prevents loading of source maps and other resources
from Windows Shares via file:// UNC URLs.

We show a warning similar to other failed source map loads, for
example when the source map was not found (404).

Screenshot: https://i.imgur.com/5wVfAO7.png

The CL also adds a setting that allows the user to disable this
behavior. Please note that the 'host' module can't have settings
themselves, otherwise we'd have a circular dependency between 'host'
and 'common'. That is why the CL places the setting in the 'network'
module, as that is also the bottle-neck for the `loadNetworkResource`
host binding. Even though the setting is in the 'network' module, the
CL places it in the sources category, as source map loading is the
dominant use case.

R=​bmeurer@chromium.org, jarin@chromium.org

Fixed: 1342722
Change-Id: I878e77592703cbfa97828650f256c2b4e4a2ac0d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3784603
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
(cherry picked from commit 1a1e7c6f50b85ad086654f9a8562efa15a943de6)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3802970
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>