ProtocolMonitor: accept various JSON formats
This includes cmd/command/method keys for the command name and parameters/params/args/arguments keys for the parameters. Currently,
we don't have a UI to give the user a hint about the right format
and those typical keys are usually used interchangeably so it looks like
we could support all of them.

Bug: none
Change-Id: I81fce8ced4b45be82c4f7c1aa32c8d3972bbcf92
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3792103
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>