DataGrid: make reading the scroll position and scrolling to happen in the same task
Currently, automatic scroll to bottom is broken (in ProtocolMonitor) if many events arrive at
the same time due to the fact that scrolling happens asynchronously
and the component might detect that the user is not at the bottom of the page
because the scroll is still being scheduled while another scroll caused
by the arriving events is happening. Replacing read/write with a call to schedule scroll using the coordinator solves the problem.

Bug: none
Change-Id: I5c4e2b9872d6bded45a85e5363ff7404cb82b186
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3791245
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>