Consistently hide call frames that match an ignore-listing hint in the console output
This CL makes the structured stack trace in the console output respect
the `x_google_ignoreList` field that was recently added to source maps.

In the Console, there are two separate bits of logic that control the
appearance of call frames in stack traces printed to the output:

1. Whether or not each call frame is displayed by default, depending on
whether it's ignore-listed.
2. When shown, whether or not each call frame is visually styled to
signify that it's ignore-listed (aka the link is "semitransparent").

Both of these now check `isUserOrSourceMapIgnoreListedUISourceCode` in
`IgnoreListManager`.

Note:  #1 happens when the UI for each call frame's row is created. #2
happens after each call frame's live location is updated. The locations
between #1 and #2 can differ, depending on whether source maps are
present. Because of this, both user-specified ignore list regular
expressions, as well as source map hints in `x_google_ignoreList`, can
match differently between row creation time (#1) and live location
update time (#2).

As a consequence, a *permanent* determination on whether to hide call
frame rows depending on a "maximum number of rows shown" constant cannot
be done during #1. Therefore, this CL also removes this logic.

Bug: 1345925
Change-Id: I0605663bed5652ce5d457ca7be87f8c4b864f41f
Signed-off-by: Victor Porof <victorporof@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3802976
Reviewed-by: Eric Leese <leese@chromium.org>