Persist memory highlight on debugger paused event
Previously, we would remove any highlight on a debugger paused event
as we could not tell if the highlighted memory object is still in
scope.
This CL implements highlight persistence by making a call to V8 after
every debugger paused event to evaluate the highlighted variable, if
one exists.

We cannot guarantee if the object returned by V8 is the same as
the currently highlighted one. We check the type and address as
a basic pre-caution. This is not bullet-proof, if we highlight a
variable X in scope A and resume the debugger until it stops in
scope B with variable Y where X and Y have the same name, type,
and address, the inspector will switch the highlight from X to Y.

Before/After: https://imgur.com/a/rvjgEfp
Design Doc: https://bit.ly/3PM7mF1
Bug: 1336568
go/extending-memory-inspector
Change-Id: I610f3c36c48778974874ae800b481cb8786f45fc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3769850
Commit-Queue: Michal Pitr <michalpitr@google.com>
Auto-Submit: Michal Pitr <michalpitr@google.com>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>