Add `nonConfigurableExperiments` and mark eye dropper experiment
Native EyeDropper is not available in some platforms
so we don't want to enable the experiment
that uses it in those platforms and we don't
want to show the experiment in the settings.

That's why I:
* Removed enabled by default for EyeDropper
experiment if the feature is not available in browser.
* Added EyeDropper experiment to
nonConfigurableExperiments.

Bug: 1348921
Change-Id: I57f4eed7da0a5a43459347301978283c067b4acc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3824817
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>