Remove WebGL support from Perf FlameChart
If you enable this, it's currently broken and does not work properly. We
could fix this, but having discussed with paulirish@ we agreed there's
no demand for this, it's realistically not something we will invest in,
so we're best to remove it.

Bug: 874116
Change-Id: I33adc1bbfab6cfa6aceb5a0ca36a2ad70b96a200
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3865458
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>