[e2e] Add simple 'Pause button' e2e test
This CL adds an e2e test that checks the CUJ of the pause button.
That is clicking it stops the script.

The test replaces an existing "browser_test" in `Tests.js` that is
flaky on linux and mac and caused issues with a V8 CL:
https://crrev.com/c/3867311

R=kimanh@chromium.org

Bug: 1246907,1289529
Change-Id: I2a639d7029c01c1b04c1a89bc027a809fcf020bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3867522
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>