Testing BFCache: wait until navigation history has caught up
When using the 'Test BFCache' button to navigate away from the page
and back, it can happen that the navigation history is lagging behind
compared to what is visible in the browser. In such cases it is
possible by repeatedly pressing the button as soon as it is active
again, to end up not at the page we started at, but instead at a page
earlier in the navigation history.

This change ensures that the browser waits until the navigation
history has caught up, before returning to the original page.

Fixed: 1317042
Change-Id: I132ffdbb6904ede128510a6cc2d1e8b870d1d90a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3890877
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>