Bugfix: Unhandled JS error resolving local names
We have error logging indicating an unhandled error reading 'sourceURL'
from an undefined value within findEntryRanges at this line. It's not
exactly clear to me why `mappings[index]` would be undefined in this
case; however, I have good data indicating that it's actually happening.
My supposition is that the `upperBound` call is returning
`mappings.length`, which is ultimately resulting in the evaluation of
`mappings[mappings.length]`, which is of course undefined.

The change I've made should be compatible with the API shape. Here, if
`mappings[index]` is in fact undefined, then the assignment of
`sourceURL` would be undefined; and consequently, the API would return
`null` in the following check.

Bug: 1368817
Change-Id: If16031d7a8ccd73285c7dc53883f3f3046b78bb0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3924034
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Auto-Submit: Robert Paveza <Rob.Paveza@microsoft.com>