[BreakpointsView] Only provide edit option for editable breakpoints.
DevTools doesn't support conditional breakpoints and logpoints for
Wasm-based languages currently, so we shouldn't provide the option
to edit breakpoints for those files (just like it's done in the
DebuggerPlugin). To maintain a somewhat easy to reason about (and
test) contract for the BreakpointsSidebarController, we mirror the
method supportsConditionalBreakpoints() on the BreakpointManager.

Bug: chromium:1346231, chromium:1371039
Change-Id: Ie1d4da8c1cd5c29876197c48964b5db526e9a420
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3937976
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>