[Accessibility] Screen reader fixes for category breakpoints
This CL ensures that child breakpoints are set to the correct aria-checked
state when a parent/category-level checkbox is clicked.

It also fixes a redundant aria-label on CheckmarkLabel components, which
was causing the label to be read twice (the checkbox is already labeled
by using the "for" attribute on the <label> element).

Bug: 1371570
Change-Id: Iddbdec785659e4236883d11d3bf95e826430e5ad
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3931906
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Leo Lee <leolee@microsoft.com>