Add `selectTextInTextNode` to TestRunner
This method comes from `Text.proto.select` which is defined in
`DOMExtensions` but is completely unused in the devtools-frontend
codebase. Therefore, this CL moves it into a utility function that we
can update the layout tests to use, before removing the method from the
Text prototype.

Bug: 1050549
Change-Id: If224c56d63ca6e6d6734a472413d7fb2f0e9b75e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3967793
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>