Header overrides editor: handle special chars in header value
Adds a test that special characters are handled correctly by the
header overrides editor.

The tab character requires special treatment: the UI shows a regular
whitespace instead of a tab. Previously when comparing the string
captured from DOM (with space instead of tab) with a string in memory
(with tab) for equality, there would always be a mismatch due to the
tab/space difference. This CL ensures that for header overrides, tabs
are always treated as spaces.

Bug: 1297533
Change-Id: I44437a11d7097a1e0c107710892d40f734914b0c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3971321
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>