Show only new headers view if header override experiment is enabled
So far, when the 'header overrides' experiment is turned on, DevTools
was showing 2 headers view tabs in the network panel in parallel. With
this CL, if the experiment is activated, only the new view is shown.
If the experiment is turned off, only the old view is shown.

Bug: 1288023
Change-Id: I16675935aa7f24af4e331d6a45e42caef0ca68ab
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4003966
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>