Prevent host bindings from loading resources from remote files
This extends the existing ad-hoc UNC blocking to all remote files.
Instead of trying to match the URL prefix, we now use the built-in URL
parser to figure out if a URL is a remote file.

The patch also changes the user visible strings to say that we are
preventing access to remote files for security reasons.

Bug: chromium:1381217
Change-Id: I73eb4d95113585b67b6e2af432942f72db364310
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4013140
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>