[sources] Simplify minification detection heuristic.
This replaces the existing minification detection, which was checking
for a line longer than 500 characters in the first 10 and last 10 lines
of a file with an easier and more reliable heuristic, which just checks
whether the average line length of the whole text is 80 characters or
more.

This correctly detects the cases we care about as minified, while not
treating user code that occassionally contains long lines (as in the
example case of https://coffee-cart.netlify.app) as minified code.

Drive-by-fix: Also properly move all the TextUtils unit tests under
a toplevel `describe('TextUtils', ...)` so that the mocha fgrep
filter works for running just these tests as expected.

Bug: chromium:1382752
Change-Id: I35e088450233c1e6f06c5da20c5c4e04e095ebf6
Fixed: chromium:1382755
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4020234
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>