[sources] Resolve source URLs consistently
Currently, there is a subtle difference in relative path resolution for
source URLs in resource script mapping and in the source map manager: If
the base URL (the URL of the containing frame) cannot be used to make
the source URL absolute (for example, if the base URL is a data URL),
resource script mapping would interpret the URL as one without a domain,
but source map manager's resolution would fail and the source map would
not be used.

With this patch, we change the source URL resolution for source maps
to be the same as for resource script mapping - if the URL cannot be
completed to an absolute path, we just keep the URL as is.

Bug: chromium:1383416
Change-Id: If2f736535b0a09603f697e4be69864ca06d6ae24
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4020558
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>