[BreakpointsSidebarPane] Allow for independent pause on exception
This CL allows non-node targets to independently select
"pause on caught" or "pause on uncaught exceptions". For node
targets (for compatibility reasons of Node LTS), we are currently
disallowing this and will lift it as soon as Node LTS has caught
up with the back-end change that allows for independent toggles
(https://crrev.com/c/4013330). For node, one can only pause on
caught exceptions if one is pausing on uncaught exceptions.

Drive-by: Only add the BreakpointsSidebarPane as listener to flavor changes if the experiment is on
Screenshot: https://imgur.com/a/7nQoaCN
Bug: 1324920, 1382762
Change-Id: Id44a6eb6d7c59870b029e9f0d3899569bbaf7e7b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4016814
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>