Scroll palette switcher panel into view
In the color picker if there are sufficiently many colors in the
palettes, opening the palette switcher causes the switcher panel to be
rendered off screen because the color picker becomes scrollable. This
change scrolls the picker to the bottom to display the switcher panel
properly.

Fixed: 1365326
Change-Id: I4c61c92c5257284d5233b1fa2ef16ee090f2e255
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4020561
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>