[l10n] Simplify fallback when locale fetching fails
We don't need to load en-US to fall back to en-US strings. After the
refactoring to "LocalizedStringSet", the fallback is already handled
there, including caching for fast future accesses.

This means the current fallback logic is no longer required and we
can simplify it a bit.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: Ifdf13bdd41db3409389adda8da3f67fff1683988
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4027205
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>