[Styles] Fix ordering of style blocks
The styles blocks in the Elements panel can sometimes be rendered out of order. The reason for this is that their insertion is scheduled as IdleCallbacks, which can be called out of order when scheduled with a timeout. This change fixes the IdleCallbackManager to make sure callbacks are processed in order by queuing them up and letting the IdleCallbacks drain the queue instead of calling the callbacks directly.

Fixed: 1382794
Change-Id: If589849bf2292c7a464b224f36afaefc710f5b5b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4020238
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>