Stub `console.clear.history` in ConsoleView test
Without this, the tests would include some logging output that
`console.clear.history` is not a valid command. It didn't fail the
tests, but it was adding noise. Let's mock out that command so that we
don't see the noise.

Bug: none
Change-Id: Ia277d412dcbc2cc9a588c23fd883776a5b145135
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4026873
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>