[Color] Add CircularColorSwatch and use it in styles and computed panes
This CL:
* Adds a new component: CircularColorSwatch
* Updates `Color` regex to also match colors with new syntax
* Updates StylePropertyTreeElement processColor to render circular color
swatch if the given color is represented in one of the new color
syntaxes[1]. This shows the circular color swatch in Styles pane
* Updates autocompletions for variables to show this color swatch for the
colors written in new color syntax
* Updates ComputedStyleWidget to render CircularColorSwatch for colors
in computed style pane.

The reason I chose for implementing a new component instead of updating
`Color` class to parse wide gamut colors is: it will require more work
and this will be tackled while implementing conversions.

So, as the next steps, what will happen is:
* We'll add conversions to different color spaces and parse colors with
new syntax to be a `Color` object.
* Remove `CircularColorSwatch` and handle rendering/interactions in
`ColorSwatch` itself.
* Implement Color Picker for the colors defined in new syntax and
remove special handling of such colors (with removing `canBeWideGamut`)

I'll tackle showing color swatch for shadow editor in a later CL.

See the screenshots for the component:
* Opaque: https://bugs.chromium.org/p/chromium/issues/attachment?aid=576179&signed_aid=PiYe3o8328N5hGz4PEsWlw==&inline=1
* Transparent: https://bugs.chromium.org/p/chromium/issues/attachment?aid=576180&signed_aid=0FkQWZ4t9410IMGDVkUjRA==&inline=1


[1]: lch, oklch, lab, oklab, color

Fixed: 1382409
Change-Id: I6adfd4a85d6552937f89708039a19c089db2811e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4013781
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>