[i18n] Generate 'locales.js' automatically from GN config
DevTools currently has a "locales.ts" file that configures all
available locales for the runtime. This configuration is duplicated
for the build step to determine what locale file to include
in the build output and what files are bundled vs fetched remotely.

This CL moves the configuration to GN and from that generates a
JavaScript file with the runtime configuration. This way we have
a single source of truth.

Follow-up CLs will clean this up further:
  - Introduce a helper template for the "node_action".
  - Handle the en-XL pseudo locale correctly in the GN files
  - Generate ES6 `Set`s instead of arrays.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: Ic43ba2fedc42f7399c81dfb37d9b12151456de4d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4027930
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>