Show the original color spec in the inspector overlay if available
To support non-rgb color specs, the renderer is sending the original
color spec along with the rgb-converted colors for the element overlay.
We need to keep the rgb-converted colors for now because they are used
for contrast computation. This CL shows the color spec in the overlay
instead of the hex values.

Bug: 1384524
Change-Id: Icfb42a0667420b47c07b3c6d5ce5bc8e9e03265c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4026875
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>