[sources] Fix location mapping for inline scripts
If the debugger stops in an inline script without source URL before the
embedding HTML arrives, the stopping position in the source view is off.
This adjusts the source positions to account for the fact that for
scripts without sourceURL, v8 provides source positions in embedder
coordinates, but the devtools maps to the embedded script locations.

Bug: chromium:1325457
Change-Id: Ia56bec7fe1c3db678c4167a0ef01d4fffbfccada
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4038144
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>