[l10n] Replace fs.writeFileSync with writeIfChanged
We want to use the functions from `collect-strings.js` and
`bake-ctc-to-lhl.js` as a build step in the upcoming CL. This means
we should write the output files only if they are changed, otherwise
we'll confuse ninja and rebuild things unnecessarily. (Ninja is
timestamp based).

This CL replaces the unconditinal write with the `writeIfChanged`
helper we use everywhere else in node scripts used during build time.

R=kimanh@chromium.org

Bug: 1185727
Change-Id: I8ebb41a2492b03292198dc4d00947792ccd5594c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4042207
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>