[l10n] Add check-strings.js script
The CL https://crrev.com/c/4043247 switches us over to use the
build output en-US.json. This means we no longer need the in-tree version
upon PRESUBMIT. We still need the original collect-strings.js to
create the en-US.ctc.json when we want to populate our translation
pipeline and to bake the locale JSON files when retrieving the
translations.

So this CL replicates a small part of the collect-strings.js script
that does the actual checking which we can use for the PRESUBMIT.

The desired end state is that we have 3 dedicated scripts for our
3 workflows: 1) Checking upon PRESUBMIT, 2) collecting the .ctc.json
when importing into the translation pipeline and 3) baking the
locale JSON files when translations are ready.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: If7a04c2b5ffe3d7f732b570c5072a888d14b16ff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4044363
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>