Add shortcut to toggle eye dropper when the color picker is open
People were asking to be able to continuously select a color
in the native EyeDroper. However, EyeDropper API is only available
after a user interaction.

So, this CL adds a shortcut for having a similar experience to
continuously picking colors with the help of a one key shortcut.

Fixed: 1392717
Change-Id: I82abbc02ea7b57fde852fc87c72b5ec9faf178ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4016819
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>