Use css parser for formatting CSS and exclude var() from property start
We were using SCSS parser for formatting CSS before which is
fixed in crrev/c/3789375 but re-introduced in crrev/c/3792110. This CL
changes that back to using CSS parser.

Currently we have a custom formatting logic in `CSSProperty` which kind
of relies on the internal details of CodeMirror. We were marking `var()`
calls as a start of a property which they are not. So, this CL excludes
them from being a property start.

I've doubts about having a custom formatting logic in `CSSProperty`
and relying on the internal details of CodeMirror. (For example,
to understand which token type is which, we need to check out the
implementation in CodeMirror side) Though for now, I don't have
a better suggestion.

We might re-evaluate why we need this formatting logic in the future
and how we can make it more robust.

Fixed: 1392813
Change-Id: I240c4dd86f0f2fc6a175fdb25e708263ae818495
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4051382
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>