[l10n] Properly re-generate en-US.json/en-XL.json
This CL properly triggers re-generation of en-US.json and en-XL.json
when any of the JS, TS or BUILD.gn files change in "front_end".

Instead of using `deps`, we use a depfile. depfiles are dynamically
created "sources" or "inputs" to an action. The collection rule writes
out all the ts/js/gn files into a depfile with Makefile format.
ninja can use the depfile to determine whether the rule needs to be
rebuilt or not. We don't want to use `deps` as that comes with
a whole host of complications around metadata/grd files that
can bubble up.

DISABLE_THIRD_PARTY_CHECK=export for re-use

Bug: 1185727
Change-Id: Ida2cfa40c17544cb46219467f410309a606cd805
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4055600
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>