Parse colors defined in new syntax as `Color` objects
Internally, we're still holding `sRGB` values which can be
out-of-sRGB-gamut and this is fine as we'll use them as:
* Transport mechanism of Colors
* Contrast calculations

It seems like contrast calculations with out-of-sRGB-gamut colors with
sRGB values are fine. So, we probably won't need to do anything specific
here. (which I'll confirm next)

This is a biiiig CL but I thought it tells a meaningful story, so
decided to keep it this way. Let me know if you would prefer seeing
changes in smaller CLs.

All the things happening in this CL, in order:
* We handle parsing new color syntaxes: `lch()`, `oklch()`, `lab()`,
`oklab()`, `color()` and create `Color` objects for them too
* Since we're able to parse these colors now, we don't need to special
case color texts before parsing and render circular color swatch where
needed. Instead, we're passing `Color` objects to `ColorSwatch` and
it decides what to render: circular color swatch or regular one.
* Removal of CircularColorSwatch component.
* Since now we're able to understand colors written in the new syntax
CSS Overview panel collects color out of the box. After talking to
Peter, I've made one adjustment to make sure those colors are
represented good (at least without overflowing to other elements
or breaking the layout).

Fixed: 1385379
Change-Id: Ic9a86addcab08a57bcba6deb5b9e433b146945c6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4057666
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>