Shared Storage: Add DevTools per-origin items view.
We add `SharedStorageItemsView`, which will be used to display the
metadata and data for a single origin's shared storage.

The data table is editable via the toolbar buttons as well as by
double clicking on the data cells to enter new values.

For the full initial Shared Storage DevTools UI prototype, see
https://crrev.com/c/devtools/devtools-frontend/+/3983292.

Bug:1372007,1218540
Change-Id: I290feb2bcd4ffe2713fe3ff23446e4c012c53b2b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4006372
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>