[A11y] Use menuitem control type for ComboBoxOfCheckBoxes headers
This CL makes sure that header items are added to the context menu as
menu items, not checkbox items, and that the screen reader announces
the item type correctly. This class is only used is the CSP Violations
view, and the header item acts as a reset button, so it doesn't make
sense for it to be a checkbox item.

Bug: 1392920
Change-Id: Ie0df4cc6dac5bcad5fbd8fad635fa5f198b2340b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4061170
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Olivia Flynn <flynnolivia@microsoft.com>