Enter keypress on new row in Storage doesn't work
In Local / Session storage data grids, you can click in to activate
editing of the last row, but you can't press <enter>. This is due to a
bug in the implementation of focus detection, in which the editor only
checks to see if the overarching data grid IS the currently-focused
element, not whether the data grid CONTAINS the currently-focused
element.

This bug blocked keyboard-only users from adding a new entry to the
Local / Session Storage tables.

Bug: 1394870
Change-Id: I433d95b5eae37632075b84326ddc1d211f13721b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4066165
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>