Link to prepopulated Attribution Reporting header validator form
The form, which is maintained by the developers of the Chromium
Attribution Reporting implementation, allows the user to see the
precise errors in their Attribution-Reporting-Register-Source,
Attribution-Reporting-Register-Trigger, or
Attribution-Reporting-Eligible headers, and can be prepopulated using
query parameters on the URL itself.

I locally verified the correct behavior of the generated links.

https://wicg.github.io/attribution-reporting-api/validate-headers

Bug: 1349866
Change-Id: I0caf5dd1921c7b69c5a2f12df3fc9281f59d333f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4076187
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>