Make elements tree ellipsis clickable to expand
The ellipsis in the elements tree panel isn't clickable now.
Giving a clear button for expanding is more in line with human intuition. And with the button in the middle, it's easier to click than the left triangle.

Screen Recording: https://imgur.com/lubIzxw

Bug: None
Change-Id: I5fbd78dff38490af08fdb01c5cd8a1ed063f3974
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3898264
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>