Shared Storage: Show origin DevTools items view even if not using API.
After feedback from some internal testers, we've decided to change the
model so that every non-opaque origin in the page will have a
`SharedStorageTreeElement` containing a `SharedStorageItemsView`
created for it.

We've also tweaked the UI strings for the metadata view to be more
descriptive, adding some additional hover text explanation for the
budget string.

Bug:1372007,1218540
Change-Id: I0af57f184e5e60e08862fb1e5021344d531d3ebf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4089404
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>