Remove focus-ring offset in styles pane
We recently added an offset to the focus-ring styles [1]. As a
consequence the focus ring in the styles pane is partly covered by
other elements and not visible.

Because this area is dense and full of information, instead of making
room for the focus ring offset, this CL removes the offset in the
styles pane and restores the previous look.

[1] https://crrev.com/c/4009038

Fixed: 1400616
Change-Id: I09340e18c5c5299c5bebafe3aa606ec63409f56f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4106709
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>