Fix disappearing icon for newly added header overrides
This fixes a bug caused by https://crrev.com/c/4110894: when adding
a new header, it would be correctly marked with the icon for
prompting the user to refresh for the changes to take effect. But as
soon as the user would edit this header's value, the icon would
disappear. This is now fixed.

Bug: 1383094
Change-Id: I0c2459e9b2c6c2e9cf80939cde952b38d7e7d3a1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4122609
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>