[Perf Panel Node] (Part 2/2) Don't skip the system nodes when build the callFrames
Currently, when we build the fake jsFrame events for cpuProfile, the system nodes ( idle node, garbage collector node & program node) are skipped. This CL just let them be handled like normal, so they will be displayed. The special case is garbage collector node, which has no parent, so we use the precious system node as its parent, which is same as JS profiler and other cpuProfile viewer.

This also solve the time calculation wrong problem, because in Perf panel frontend, it will auto fill the empty gap, so for example when there is no '(idle)' node, the previous function will fill the time period of '(idle)', and Perf panel also will also add this time to the previous event, so the previous function's time is amplified.

Screenshot: https://imgur.com/a/QMFyWUr
Fixed: 1382733, 1080918
Change-Id: I2f741aed7a32ab3971213605168a81857e27ede8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4118545
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>