[Perf Panel Node] (Part 1/2) Use root event instead of programEvents & functionEvents
We don't need programEvent and functionEvent anymore. The reason is the programEvent is created whenever we find a node that is not '(idle)', and the functionEvent is created when the node is neither '(idle)' nor '(program)'.
This is also the reason why (garbage collection) and other events are missing, because they are suppressed by the functionEvent.

Instead, just simply append a (root) event, which is same time range with the whole profile. so all other nodes can be the displayed inside its time range. And the root event itself won't be displayed because it has no child.
Also another benefit is the root event will occupy the time gap between start time and first sample, so the Performance panel won't truncate this time period.

Screenshot: https://imgur.com/a/QMFyWUr
Bug: 1382733
Change-Id: Id61d424ba023746e3b6e341a5fab4aa83a6d2156
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4118739
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>