[sdk] Remove unsound caching from the SourceMapManager.
The SourceMapManager was trying to cache sourcemaps by their "source map
id", which was basically the concatenation of sourceURL (of the client)
and sourceMappingURL (of the sourcemap). This caching is unsound by
itself, since the content of both the client and the sourcemap can
change without changing their URLs. So there was already quite a bit of
ducktape on top (in DebuggerModel) to forcibly detach sourcemaps when
there it would detect a clash (based on execution ID, sourceURL and
sourceMappingURL). But even this ducktape still left plenty of ways open
in which a stale sourcemap could be fetched from this cache (one of the
easy ways to reproduce is demonstrated in https://crbug.com/1403362).

One alternative how to go about this would be to also include the client
content hash into the mix (i.e. the script hash for JavaScript). But
right now we don't have the hash available for the stylesheet contents,
and even if we did, it wouldn't fix the fundmental problem either, since
the sourcemap (and sources) can change while still producing the exact
same bundle in case of both JavaScript and CSS.

So the appropriate solution is to remove this caching layer and instead
rely on Chrome's caching for sourcemaps (parsing sourcemaps is fairly
cheap on common developer hardware, so that's a complete non-issue).
This CL thus drastically simplifies the SourceMapManager and adds a
significant chunk of test coverage for the interesting behaviors.
Specifically this means now that there's a 1:1 mapping between
sourcemaps and clients, rather than this strange 1:N mapping that we
used to have before (which would almost never trigger except when it
shouldn't).

Drive-by-fix: Remove i18n from the SourceMapManager, as this is a
layering violation and we generally don't localize internal error
messages (like the warn() call in SourceMap.ts is not localized either).

Fixed: chromium:508270, chromium:1403362
Change-Id: I2c7d29635ba085c433bdb3cc691f720e9c158808
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4120261
Reviewed-by: Yang Guo <yangguo@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>