[console] Properly toggle message separators.
When displaying the result of a command right next to it, we don't
render a separator in between the two "console messages". However,
we didn't update the state of the separator as the filter status
changed. This CL fixes that.

Fixed: chromium:1402092
Change-Id: Ie7d6481893d111e57375962629ed4ce74f07dfdf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4116526
Reviewed-by: Yang Guo <yangguo@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>