[cleanup] Properly update live locations of inline scripts and styles.
The fix in https://crrev.com/c/2742490 put the logic for updating the
live locations of inline scripts into the wrong place, and also did not
perform the live location update for inline styles. It was the wrong
place because the CompilerScriptMapping is concerned with sourcemap
related UISourceCodes, whereas the ResourceMapping is responsible for
documents (with inline scripts and styles). The previous fix worked by
chance due to a specific order of events for scripts.

This patch puts the logic into the correct spot, and extends it to also
update the locations for inline styles once the document is loaded.

Bug: chromium:1184162, chromium:1403432
Change-Id: I659bc8899b8b2cec03f9fd225648c6d305194694
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4127779
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>