[cleanup] Remove sourcemap caching work-around from DebuggerModel.
Following up on https://crrev.com/c/4120261, we no longer perform
unsound caching inside the SourceMapManager, but instead we now have a
1:1 mapping between Script and SourceMap. Thus we don't need to worry
about stale sourcemaps anymore when scripts are reloaded, and we can
remove the sourcemap caching work-around that was put in place in the
DebuggerModel, which would forcibly detach prior sourcemaps when it
detects a situation that could lead to a stale sourcemap otherwise.

This greatly simplifies the DebuggerModel wrt. sourcemaps and makes
it much easier to reason about what's going on.

Bug: chromium:1403362, chromium:1403432, chromium:508270
Change-Id: Ic52bdf89f6aa18ccdd5d44d37f3925f91ebb7e55
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4128463
Commit-Queue: Victor Porof <victorporof@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Victor Porof <victorporof@chromium.org>