Shared Storage: Add reset budget button and align info icon.
We add a button to the `SharedStorageMetdataView` to reset the entropy
budget. This will make it easier for developers who are testing; if
they exhaust the budget, they won't have to wait for it to be reset
or use workarounds like clearing browsing data or logging in from a
different profile.

https://imgur.com/a/77QAol0

Bug:1372007,1218540,1403414
Change-Id: I198c3957dbb7877f3958d63983b19ce921831c0b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4098470
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Peter Müller <petermueller@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>