Make LayoutShiftsHandler return source not Image element
An `HTMLImageElement` cannot be passed from a worker back to the main
thread. Therefore in preparation for running this code in a worker we
need to not generate the image in the handler but instead just return
the source. When we then use this data, we can construct the image in
the main thread.

Bug: 1399473
Change-Id: I3f33a775c356ad1b17188d673b05f33546ebbf64
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4154494
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>