[bindings] Refactor computation of mapped lines for sources.
This introduces a new `getMappedLines()` method on the debugger
workspace binding, which consults the appropriate methods in the
debugger plugins, as well as the mappings, to compute the set of
lines that map to scripts (for the breakpoint line gutter).

With this CL we now properly indicate the script lines within
HTML documents, and we also remove one of the two last remaining
static methods on the `CompilerScriptMapping`.

Note that even with this refactoring, we are still lying about
breakable lines in the breakpoint gutter, since just because a
line is mapped to a script doesn't necessarily mean that it's
possible to set a breakpoint on this line. Fixing this will need
a combination of the mapped lines information with information
about possible breakpoints from V8.

Bug: chromium:1185770
Change-Id: Ie8481df799d3986a24caea5117a884c2a1cd0433
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4161371
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>