[cleanup] Simplify DefaultScriptMapping.
Remove the global WeakMaps from the DefaultScriptMapping and replace
them with per-instance Maps that are properly updated. Since there's
still the static `scriptForUISourceCode()` method around, we use the
same temporary workaround as in the `CompilerScriptMapping` of having
a global Set of all the instances meanwhile.

Also most importantly the UISourceCode to Script mapping is always
a 1:1 reverse mapping of the Script to UISourceCode mapping (while
our implementation was maintaining a Set of scripts for each of the
UISourceCodes in the global mapping).

With this change it's not only more efficient (avoiding the potentially
huge WeakMaps and favoring much cheaper - from a GC perspective - Maps),
but also easier to reason about the behavior of this mapping.

Bug: chromium:1379221
Change-Id: I5ed2493d2a001f315d690b7883a1d68ce369b1e5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4161377
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>