[cleanup] Remove the leaky `SourceMap` interface abstraction.
Previously we had an interface `SourceMap` and a concrete implementation
`TextSourceMap`, but this abstraction was already leaky and in a few
places we even casted directly to `TextSourceMap`. There's not really a
benefit in maintaining this abstraction and generally we avoid the
"programming to interfaces"-paradigm, so let's just remove the SourceMap
interface and be honest with ourselves here.

Bug: chromium:1262113
Change-Id: I67bf0484f55b16ef9bc72119f3dd5d4c85c63968
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4161382
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>