[Colors] Allow out-of-gamut colors & clipping on conversion and parsing
When colors are parsed or converted, this changes adds clipping behavior and keeps track of out-of-gamut colors and "benign" clipping. For example, passing a lightness greater than 100 to lab() does not produce an out-of-gamut value, but the lightness is still clipped.

Drive-by: Make channels public on the new color classes.

Bug: 1395782
Change-Id: I2abf9df9cb3a9e72ef30cca8c748004194e36691
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4091627
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>