Move TraceEngine parsing into a worker
This CL moves the new trace engine into a worker.

The `Model` is kept on the main thread, but the processor that does the
work of running trace events through the handler now is run in a worker.
The model hides this implementation detail and communicates to/from the
worker using messages.

This CL also updates the panel to call into the new engine. It does this
without `awaiting`, so we won't materially slow down the user
experience, and it does not yet use the data in any form. This is just
the first step towards beginning to use the data to power the panel.

Fixed: 1399473
Change-Id: I50aae988d88d7fc916a9400ae81a8808c02f098e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4161378
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>