[layout tests] Throw error if the wrong breakpoint sidebar pane is used
The breakpoint sidebar dumping and awaiting functions are only to be
used with the old breakpoint sidebar pane that is soon going to be
replaced. Make sure to throw an error to make this clear, in case
tests are changed to make use of this functionality.

Bug: 1407586
Change-Id: I86fd50eab52ba4404502c0665b00047195ca2da5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4171280
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>