Show icon for active header override in editor tab header
Displays a file icon with a purple dot in the tab header for the file
editor of an 'active' header file. A header file is active if it
matches (= potentially contains header overrides for) some of the
active page's requests.

The NetworkPersistenceManager emits events to notify the open editors
to update their tab headers. When updating the tab headers query the
NetworkPersistenceManager whether their file is active or not.

Bug: 1408136, 1313436
Change-Id: Iebdc0171451829ae21e9edaf2f554d83925cc88c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4161376
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>