[cleanup] Remove BreakpointEditDialog#isLogpoint property
We only need the property to set the initial value for the <select>
and once we are done to determine whether it's now a logpoint or not.

For this we don't need to maintain the property as we can check
the combobox value directly when we close the editor.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: Ic2f9b80c8c1cb586bb8f5107271396501fd66b08
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4194328
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>