[BreakpointView] Show differentiating path for ambiguous file names
For ambiguous file names, we now show a part of the path to that
file that helps to differentiate the breakpoint groups next to
the file name.

When differentiating ambiguous files, we try to find a part of the
url that is uniquely identifying a path in terms of the other
files, if possible. From that unique path, we only show the right
most folder name. If we cannot find such a unique path, we show the
whole path instead.

This happens e.g. for React, where it is common to have an index.js
file for each component.

Before: https://imgur.com/OEUR6bV
After: https://imgur.com/cLneY8H

Fixed: 1403924
Change-Id: Ib9fb822ada120af36618ac76d2c9e524f1c8435b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4196916
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
(cherry picked from commit 0801f17e15d0ef3bf9fdc473fd114c66c66aa456)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4227692