[unittest] Add test for file system persistence
This replaces the persistence-move-breakpoints layout test.
The test checks if 1. breakpoints get moved from the file system
uiSourceCode to the network uiSourceCode as soon as a binding is
created, and 2. whether they get copied back to the file system
uiSourceCode as soon as the binding is removed.

Bug: 1407586
Change-Id: I6acfb5b14bd26e0ba2bbd954c94ebd942383680c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4203270
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>