[console] Console message anchor links to breakpoint/logpoint
Conditional breakpoints and logpoints now have special sourceURL
annotations. We can use these "marker sourceURLs" to distinguish
console messages that originate from breakpoint conditions
from user console messages by looking for these markers in the
stack trace.

This CL adds the necessary logic to `ConsoleViewMessage`. Please note
that just checking the top-most frame for the marker sourceURL is not
enough: Breakpoint conditions can create and leak closures so there
could be arbitrary frames on-top of the actual pause frame.

As a heuristic we search for the last marker frame on the stack
and take its parents. In most cases this should be the pause frame
and where the breakpoint actually lives.

R=jarin@chromium.org, kimanh@chromium.org

Bug: 1027458
Change-Id: Ib839b8267575ee1177a397d8460970823cacb338
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4209271
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>