Add class for lazy loading CSS property documentation
This CL adds a small wrapper class that is reponsible for fetching
the @vscode/web-custom-data CSS info.

Creating a new instance immediatly triggers the fetch and a small
helper method allows looking up info by CSS property name.

Note that we intentionally don't return any promise to the data
from the lookup method. We want to show the info on hover and if
we return a promise we have to explicity handle the case where the
promise takes too long to resolve.

We'll wire up the data loaded here with the small lit component
introduced in https://crrev.com/c/4183651.

R=ergunsh@chromium.org

Bug: 1401107
Change-Id: Ia09d87036ac9d0d28f328a3501f8cc452812b3b3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4224091
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>