Output Karma text coverage to file
It is useful to have the Karma text output, but the output for me is
much easier to parse in a file, rather than in the command line. This
CL changes the behaviour so that the text coverage is written to
`karma-coverage/coverage.txt`.

Bug: none
Change-Id: Iaacb1fd421e887fe49f690a0374c01b5b47eaad0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4234964
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>