[unittest] Add test for setting breakpoints in pre-added targets
This CL adds two tests that capture the current behavior of setting
breakpoints on pre-added targets that share the same script.

The breakpoints should correctly resolve in both targets, but
currently that is not the case and is dependent on a race:
depending on which target is processed first, one of the targets
may set the breakpoint on the incorrect script. The reason is
that we do not add all uiSourceCodes from all targets to the
breakpoint, and thus, not all locations are correctly resolved and
we make use of a fallback. More detailed
information on the race is included in the bug description.

These two tests need to be adapted as soon as the bug is fixed.

Bug: 1415258
Change-Id: If4f711b47b1a6ce06ead57e310a4404517bc2233
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4241364
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>