[console] Prefer linking to breakpoint edit dialog over scriptId
This CL switches around the console message anchor linkification.
With this CL we prefer to link to edit breakpoint dialog if we see
the special marker URLs over prefering to link to a scriptId.

This enables the use case where we get Runtime.exceptionThrown
messages from breakpoint conditions where we also want to link to
the breakpoint and not the hidden VM script.

This CL also does a small refactoring: We hoist the RuntimeModel
check one layer higher. This allows us to inline all of the other
methods directly into 'buildMessageAnchor' getting rid of one level
of indirection.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: I7aa27d787d2970ff913e8e6189805dc9078e9c2f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4253690
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>