Fix ESLint custom names not respecting custom Event class
This ESLint rule assumes that if a class extends an Event class, it is
extending the built in class and applies checks accordingly. However, if
we are in code that has a local Event class defined, then we should not
apply this check.

Bug: none
Change-Id: I1ce94299074c44812fd2845fe7e487a2c14b140c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4259791
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>