Re-enable flaky script-in-multiple-workers_test
This test was suspected to be flaky due to a race in the
BreakpointManager, in which we only set a breakpoint for the
primary uiSourceCode (https://crrev.com/c/4248192/13). Since this is
fixed, we can now try to re-enable the test.

Fixed: 1368493
Change-Id: Iaa844187c450526be78fda26b1c92ea84a0041a9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4260209
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>