Automatically reveal '.headers' file in sidebar
When clicking on the 'Header Overrides' link in the network panel,
this click takes the user to the Sources panel and opens an editor for
the corresponding '.headers' file. With this CL, the same file will
also be selected in the sources tree in the sidebar.

Bug: 4255549
Change-Id: I50ed62b731e342819d618ef52b8d3627e0a71f4a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4255549
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>