[sources] Fix resource origins for sourcemap URL clashes.
Previously we'd refer to resource origins by URL, but that doesn't work
in case of URL clashes. With https://crrev.com/c/4183652 (the removal of
the '? [sm]' suffix) we made these clashes more prominent (for example
when using the default tooling for Vue or Svelte). So instead of going
by URL, reveal the actual UISourceCode directly instead.

Bug: chromium:1408033
Change-Id: Ic15788f1fd1c3323fbbc1a4557a9f49ebd9530a9
Fixed: chromium:1416912
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4262670
Reviewed-by: Simon Zünd <szuend@chromium.org>