Remove bind_id from EventPayload
The EventPayload type declares it as required, but recent traces we
record do not have this field. Perhaps it used to exist, or does in
rare circumstances, but grepping for bind_id in the codebase reveals
that we do not actually ever use it. We parse it, and store it on the
Event class, but it is never used or shown anywhere.

Fixed: 1417590
Change-Id: I2547470c8362cc3ef54edac82a7c2124512f0d93
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4271290
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>