[BreakpointManager] Don't set a breakpoint if no uiSourceCode is valid
Starting with crrev.com/c/4248192 we look through all *other*
targets and also add the compatible uiSourceCodes. However, we still
treated the primary uiSourceCode special, as in: if this uiSourceCode
is invalid we would still try to set a breakpoint on it.

With this change, we don't treat the primary uiSourceCode differently
than other uiSourceCodes of other targets. We look for all compatible
uiSourceCodes through all targets, and if no valid uiSourceCode
can be found, we do not set any breakpoint.

Bug: 1415258
Change-Id: I0a4c4f73b08499d411f1978af6c4f0370dab2427
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4274088
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>