Move 'ConsoleHistoryManager' into 'text_editor' component
We plan to share the implementation between console and breakpoint
edit dialog so as a first step we move the class into the
'text_editor' component and rename it to 'AutocompleteHistory'.

Note that the implementation is identical, the only difference is
that we use private fields/methods instead of the 'private' keyword.

The CL also adds the first set of unit tests. A follow-up CL will
move the backing setting directly into the 'AutocompleteHistory'
class.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: I6d6e16974860d9a511f23a1da4a95bedee3f6c98
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4274090
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>