Let 'AutocompleteHistory' manage the history setting
This CL changes the AutocompleteHistory class to take the actual
backing setting as a parameter at construction time.

This means users of the class will no longer have to manually sync
the setting with the underlying data, but the history class
actually manages that.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: I8e95c47d5ac84a401f1be32422e965c0f16f7249
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4274346
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>