Fix focusing highlight in DataGrid for edited items
The root cause is this:
- When we open the InplaceEditor, we remember the old tabIndex value and
  then set tabIndex to 0.
- When we are done with editing, we restore the old tabIndex value.
- We use `element.tabIndex` to get the old tabIndex value. The problem
  is that when the tabIndex attribute is not defined, it defaults to 0
  or -1 depending on the element type. For td this is -1.

So when we edit entries in a DataGrid, they always end up with a
tabIndex of -1, which makes then not focusable, and therefore the
CSS rule for blue highlight does not apply.

The fix is to use element.getAttribute('tabIndex') instead.

R=bmeurer@chromium.org

Fixed:1340842
Change-Id: Ie43217f81b3ed3fd8aeb16963df3b57981b81638
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4274094
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Yang Guo <yangguo@chromium.org>