Performance: Drop ranges from TimelineModel
This has bugs (see attached) where a lot of the trace file is cut off
and not shown. Broadly we do not think the idea of ranges is useful, we
typically want to show everything that occured in the trace, so let's
remove them. The new trace engine has no concept of ranges either, so
this will mean we do not have to implement them in that engine.

Change-Id: I60debad02b0f1bfbac123b9c3a1172d95324fc02
Bug: 1359507
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3862704
Auto-Submit: Paul Irish <paulirish@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>