Move historyCompletions into TextEditorHistory class
The job of the method is it to take a current CodeMirror state,
a text position where autocompletion was requested, and whether the
user explicitly requested completions; and return a suitable list
of suggestions.

This CL moves the code into 'TextEditorHistory' and changes it
slightly to only depend on the input context.

The CL adds unit tests for the main behaviors.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: Ife6988f06eec7429844016f5a1e32a46b101cfb4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4278845
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>