[RPP] Properly group raster threads in timeline
A regression was introduced in https://crrev.com/c/4271788 where the
count of raster threads was not properly tracked. As a consquence,
one track was rendered for each raster thread, when in fact we
want to group raster threads in a single track.

See image diffs on https://crrev.com/c/4286875 for an example of the
regression.

Bug: none
Change-Id: If3106e513758ad667bbe1f89bf4ac5459111fe7e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4295033
Auto-Submit: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>