Remove execution of new trace engine
It has implications on the time taken for DevTools to process a trace,
and right now on ToT we are not using the data from the parsing. Let's
remove it for now, and bring it back when we are using the data.

Bug: 1420240
Change-Id: I80b1233f2d44102bf8b6a1ef41f070157c872a20
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4295978
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>