[Header overrides] Prevent unset headers from being incorrectly marked as being overridden
If a non-existing header is causing a request to be blocked, DevTools
shows the header in the network panel in order to attach an error
message in the UI.

Such unset headers currently show up as being overridden. This CL
fixes this.

Bug: 1420283
Change-Id: I67f822285b6b90745ca7d14b742255ba25cde449
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4295181
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>