[Header overrides] Abort editing of unset headers should not mark header as overridden
When editing an unset header, it is correctly marked as being
overridden as soon as it is being edited. But when aborting the edit
by pressing 'Escape', the header is incorrectly still marked as being
overridden. This is fixed by this CL.

Fixed: 1420283
Change-Id: I4374124627389908499c065bff7c3a334db6f74a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4296545
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>