[RPP] Add timings to legacy timings track in TimelineModel
This is currently done in the data provider when it should be done in
the same place where other tracks are built (TimelineModel). Because we
will tear the data provider appart in crbug.com/1416533, it would be
convenient to define what should be part of the new system and what
should be kept in the legacy system. This piece of code should be kept
unfortunately, so this moves it to a file that will remain untouched.

The reason we need to keep this is because the track's data is used
outside the flamechart, in the details pane. So we need to make sure it
is properly built. Once the timings track in the flame chart is migrated
to the new system, we will have two pieces of code that look like
duplicates in that they both add timings to timings track. The
difference is one is done for the flame chart and the other for
compatibility with other features. When this happens I'll add an
explanatory comment to the legacy code remarking the duplicate code.

Bug: 1416533
Change-Id: Ia898bfc402c7dfc3d91984917c3daaa29d3db4fa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4307569
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>