Run mac_rel builder in CQ
To prevent the CI builder from turning red after updates to screenshots on other platforms, we require this CQ builder run and to collect golden screenshots.

Furthermore, since the builder appears to be more stable, we plan to transition to using it for blocking CQ and closing the tree with the corresponding CI builder.

Bug: chromium:1413723, chromium:1422180
Change-Id: I647f123ccb1b1b2cc055f1a502f291d9a32d5ed2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4314920
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>