[RPP] Use the new engine in the Timings track
NOTE: This has to be landed *after* the new engine has been reenabled

The first track appender is introduced to add the Timings track into the flame chart data (TimingsTrackAppender).

This appender consists of the bits previously in the TimelineFlameChartDataProvider that were in charge of appending the data of the track along with the bits that provided extra feature for events in the track (like stylings and display names).

All these parts are moved to a common place to handle the details of the track together in the same module, but separate from other tracks. The source of data used by the track appender is the output of the new engine

In the future further test coverage will be added using dedicated screenshot tests.

Bug: 1409044
Change-Id: I007d5ffd93e310cba26546fa09829e02e3cc581c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4295982
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>