Emit PrimaryPageChanged event only when the primary page actually changes
Do not emit the event on navigation of non-primary outermost frames
(e.g. main frame of prerender target).

Do not emit the event on navigation of non-outermost main frames
(e.g. OOPIF main frames). (It is ok to make this change, the listeners
to the `PrimaryPageChanged` event have been audited and do not react
to receiving the event from non-outermost main frames -> some listeners
can be simplified in follow-up CLs because the check for
`primaryPageTarget` will be redundant)

Bug: 1393057, 1422612
Change-Id: Ia2d307e9a605dd12f5dd32b360cc2f7de2629ce4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4320937
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>