Only resume worker sessions that are waiting for the debugger
Resuming a worker that is not waiting on the debugger will resume the
worker even if there is another worker session still waiting on the
debugger.

This CL is a workaround to prevent DevTools from prematurely resuming
worker targets.

This was causing issues for Lighthouse listening to network requests:
https://github.com/GoogleChrome/lighthouse/pull/14212

Bug: 1423096
Change-Id: I757f4fd3ee0969ea33e1acc95e1190c99146a023
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4326967
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>