Fix web vitals overlay not appearing
The issue was that the import for the component was a type only import,
so it was dropped at build time. We fix this by ensuring we refer to the
component via a property on it, thus ensuring that the import is
required and not dropped.

Fixed: 1423967
Change-Id: I41aa953d05944301e67b36d364cf542d026d0284
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4337915
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>