Fix NameResolver#resolveExpression for line-by-line source maps
The sources panel uses 'resolveExpression' to figure out what JS
expression to send to V8 when the user hovers over text while
paused.

The CL https://crrev.com/c/4196920 changed the behavior for 'cheap
source maps' (only line-by-line mappings) as produced by webpack
in dev mode. Instead of resolving to an empty string, the CL
produced the whole line instead, leading to a broken popover.

This CL fixes the popover by doing some additional validation.
Instead of only doing a reverse mapping from:

  "hovered text"  --->  "compiled text"

We do an additional "re-mapping" from

  "compiled text" ---> "authored text"

and check that "hovered text" and "authored text" have the same
prefix modulo whitespace and punctuation.

This again leads 'resolveExpression' to resolve to an empty
string for more complex lines, in which case we'll correctly
fall back to "hovered text". This works very well for webpack
in dev mode, as the bundle is not minimized, so the hovered
expression matches the expression in the compiled bundle.

The regression test for this CL is added on the Chromium side:
https://crrev.com/c/4337509

R=bmeurer@chromium.org

Fixed: 1422883
Change-Id: I52d058c5232b5aa3b6e091d3eb01c55d42b780f1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4338142
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>