Clean up Trace Engine update events
From a previous world where we thought we might have multiple engines at
once, we had a system of events that werre quite confusing. I have
removed these and replaced it with two events: one for progress on the
model parsing, and one that is emitted when the parsing is fully
complete.

I also tidied up some accidental duplication with the trace processor
events that it emits on updates, and added a test for the behaviour too.

Bug: none
Change-Id: I54ca2a8ad596732dbfdda14eb8c43abc912ded6a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4336482
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>