[deps] Use Chromium's node binary instead of the DevTools one
Currently roll_deps.py only works in a standalone DevTools checkout.
That is because we resolve the Node.js binary in the DevTools
directory. In a Chromium checkout that binary does not exist as
we don't do transitive deps for `gclient sync`.

This CL changes roll_deps.py to use the Node.js binary of the
Chromium repository instead. It's guarnateed to be there since we
run `gclient sync` in the Chromium directory as part of the script.

So now the script works both with a standalone and an integrated
checkout.

R=alexrudenko@chromium.org

Bug: None
Change-Id: I39e7f0d27dad1d1c820ebeb9261619b57a65ae0f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4341504
Auto-Submit: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>