[Accessibility] Update aria label to color picker in layout pane
When using 'Tab' to navigate to the color picker element in layout tab of Elements panel, the screen reader announce "Choose the overlay color for this element" , which didn't include the color picker
 detailed information. The root cause is that the <input type="color"> wrapped in a <label> element that takes focus instead.

This fix update the focus of <input> and <label> element with the correct aria-label. And
Screen reader announce "Choose the overlay color for this element (with detailed color info)" instead.

Bug: 1425199
Change-Id: I019dedef0b353034c206f18fbf059df08906cd4e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4348840
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Yanling Wang <yanlingwang@microsoft.com>