[DevToolsIcons] Rename --color-primary to --color-primary-old
We want to introduce GM3 --color-primary to use in icons. Currently
there is a conflict with the already existing --color-primary.
At the moment we don't want to change the color everywhere, so we need
to rename the old one. Later we will get rid of the --color-primary-old.

Bug: 1210894
Change-Id: I2871b41a022f438f019ab75482f4444f541c5bcf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4353012
Commit-Queue: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>