Mark first primary page target as scope target by default.
This is mostly to simplify unittests that are not concerned with the
scoping but without this need to set a scope target in each test case.

It is also arguably less surprising API overall, as without this
registering a scoped observer doesn't have any effect unless scope is
set.

Bug: 1417112
Change-Id: I9fb780ef9ab38edb590668e6a34667705975d595
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4355376
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>