Update goldens even if threshold is met when force update flag is set
Before this CL setting the flag meant that screenshots would be updated
only if the diff that was obtained in a test was above the defined
threshold. So that screenshots are more likely to be kept updated this
CL changes this behaviour so that if the flag is set, any image diff >0
would trigger an update.

Bug: none
Change-Id: Ic7af68589d09dd6082aeb5237b3dbb9c4fd138f6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4370889
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>